inventory-converter/requirements.txt
anyio==3.6.2
appdirs==1.4.4
attrs==22.1.0
cattrs==22.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
exceptiongroup==1.0.4
fastapi==0.87.0
h11==0.14.0
httptools==0.5.0
idna==3.4
oauthlib==3.2.2
pyactiveresource==2.2.2
pydantic==1.10.2
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python3-discogs-client==2.6
PyYAML==6.0
requests==2.28.1
requests-cache==0.9.7
ShopifyAPI==12.2.0
six==1.16.0
sniffio==1.3.0
starlette==0.21.0
typing_extensions==4.4.0
url-normalize==1.4.3
urllib3==1.26.13
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
